Rename LineChart to Time Series
Recharts requires specific set-up to support time series; we don't yet
need to support other line chart displays.
