+++ src/renderer/components/DummyDashboardFragment.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import { BarChart, LineChart, PieChart } from '../components';
+import { BarChart, TimeSeriesChart, PieChart } from '../components';
 import { barData, pieData, lineData } from '../utils/dummy_data';
 
 const DummyDashboardFragment = ({ className }) => (
@@ -9,7 +9,7 @@ const DummyDashboardFragment = ({ className }) => (
     <div className={`${className}__panel ${className}__panel-mock`}><BarChart data={barData} dataKeys={['pv', 'uv']} /></div>
     <div className={`${className}__panel ${className}__panel-mock`}><PieChart data={pieData} /></div>
     <div className={`${className}__panel ${className}__panel-mock`}>
-      <LineChart data={lineData} dataKeys={['value', 'other']} />
+      <TimeSeriesChart data={lineData} dataKeys={['value', 'other']} />
     </div>
   </React.Fragment>
 );
