+++ src/renderer/components/charts/__tests__/TimeSeriesChart-test.js
@@ -1,11 +1,11 @@
 /* eslint-env jest */
 import React from 'react';
 import { shallow } from 'enzyme';
-import LineChart from '../LineChart';
+import TimeSeriesChart from '../TimeSeriesChart';
 
-describe('<LineChart />', () => {
+describe('<TimeSeriesChart />', () => {
   it('defines a Line series', () => {
-    const wrapper = shallow(<LineChart data={[{ value: 1 }]} dataKeys={['value']} />);
+    const wrapper = shallow(<TimeSeriesChart data={[{ value: 1 }]} dataKeys={['value']} />);
     expect(wrapper.find('Line').length).toBeGreaterThanOrEqual(1);
   });
 });
